+++ src/program-instructions.ts
@@ -202,9 +202,20 @@ export async function placeOrderIx(
   side: Side,
   marginAccount: PublicKey,
   authority: PublicKey,
-  openOrders: PublicKey
+  openOrders: PublicKey,
+  whitelistTradingFeesAccount: PublicKey | undefined
 ): Promise<TransactionInstruction> {
   let marketData = Exchange.markets.markets[marketIndex];
+  let remainingAccounts =
+    whitelistTradingFeesAccount !== undefined
+      ? [
+          {
+            pubkey: whitelistTradingFeesAccount,
+            isSigner: false,
+            isWritable: false,
+          },
+        ]
+      : [];
   return Exchange.program.instruction.placeOrder(
     new anchor.BN(price),
     new anchor.BN(size),
@@ -238,6 +249,7 @@ export async function placeOrderIx(
         oracle: Exchange.zetaGroup.oracle,
         marketNode: Exchange.greeks.nodeKeys[marketIndex],
       },
+      remainingAccounts,
     }
   );
 }
@@ -704,4 +716,7 @@ export type StateParams = {
   readonly strikeInitializationThresholdSeconds: number;
   readonly pricingFrequencySeconds: number;
   readonly insuranceVaultLiquidationPercentage: number;
+  readonly nativeTradeFeePercentage: anchor.BN;
+  readonly nativeUnderlyingFeePercentage: anchor.BN;
+  readonly nativeWhitelistUnderlyingFeePercentage: anchor.BN;
 };